moving models to a separate folder
