+++ Common/Common.csproj
@@ -160,9 +160,12 @@
     <Compile Include="TempDirectory.cs" />
     <Compile Include="VisualStudioProjectParser.cs" />
     <Compile Include="YamlParsers\Models\ConfigurationLine.cs" />
+    <Compile Include="YamlParsers\Models\DepLine.cs" />
     <Compile Include="YamlParsers\Models\ModuleConfiguration.cs" />
     <Compile Include="YamlParsers\Models\ModuleDefaults.cs" />
     <Compile Include="YamlParsers\Models\ModuleDefinition.cs" />
+    <Compile Include="YamlParsers\Models\ParsedDepsSection.cs" />
+    <Compile Include="YamlParsers\Models\ParseDepsSectionResult.cs" />
     <Compile Include="YamlParsers\Models\YamlInstallSections.cs" />
     <Compile Include="YamlParsers\ModuleYamlFile.cs" />
     <Compile Include="YamlParsers\SettingsYamlParser.cs" />
