+++ Common/YamlParsers/Models/DepLine.cs
@@ -1,0 +1,36 @@
+using System;
+using System.Diagnostics;
+
+namespace Common.YamlParsers.Models
+{
+    // [DebuggerDisplay("{" + nameof(DebuggerDisplay) + ",nq}")]
+    public class DepLine: IEquatable<DepLine>
+    {
+        public DepLine(Dep dependency): this(false, dependency)
+        {
+        }
+
+        public DepLine(bool isRemoved, Dep dependency)
+        {
+            IsRemoved = isRemoved;
+            Dependency = dependency;
+        }
+
+        private string DebuggerDisplay => IsRemoved ? "-" + Dependency : Dependency.ToString();
+
+        public bool IsRemoved { get; set; }
+        public Dep Dependency { get; set; }
+
+        public bool Equals(DepLine other)
+        {
+            if (ReferenceEquals(null, other))
+                return false;
+            if (ReferenceEquals(this, other))
+                return true;
+            return IsRemoved == other.IsRemoved && Dependency.Equals(other.Dependency);
+        }
+
+        public override string ToString() => DebuggerDisplay;
+
+    }
+}
