+++ Common/YamlParsers/Models/ParseDepsSectionResult.cs
@@ -1,0 +1,8 @@
+namespace Common.YamlParsers.Models
+{
+    public class ParseDepsSectionResult
+    {
+        public DepsContent ResultingDeps { get; set; }
+        public ParsedDepsSection RawSection { get; set; }
+    }
+}
